+++ src/engine/TileMap/IsometricEntityComponent.ts
@@ -1,6 +1,13 @@
 import { Component } from '../EntityComponentSystem/Component';
 import { IsometricMap } from './IsometricMap';
 
+export interface IsometricEntityComponentOptions {
+  columns: number;
+  rows: number;
+  tileWidth: number;
+  tileHeight: number;
+}
+
 export class IsometricEntityComponent extends Component<'ex.isometricentity'> {
   public readonly type = 'ex.isometricentity';
   /**
@@ -8,14 +15,20 @@ export class IsometricEntityComponent extends Component<'ex.isometricentity'> {
    */
   public elevation: number = 0;
 
-  public map: IsometricMap;
+  public readonly columns: number;
+  public readonly rows: number;
+  public readonly tileWidth: number;
+  public readonly tileHeight: number;
 
   /**
    * Specify the isometric map to use to position this entity's z-index
    * @param map
    */
-  constructor(map: IsometricMap) {
+  constructor(mapOrOptions: IsometricMap | IsometricEntityComponentOptions) {
     super();
-    this.map = map;
+    this.columns = mapOrOptions.columns;
+    this.rows = mapOrOptions.rows;
+    this.tileWidth = mapOrOptions.tileWidth
+    this.tileHeight = mapOrOptions.tileHeight; 
   }
 }